+++ CHANGELOG.md
@@ -7,6 +7,11 @@
   - `addBreakOperandFP`
   - `addReturnHandlerFP`
   - `addCallHandlerFP`
+- add new setCallback methods:
+  - `setupMemoryCallback` (split from setupCallback)
+  - `setupDeviceCallback` (split from setupCallback)
+  - `setupMemoryCallbackFP` (split from setupCallbackFP)
+  - `setupDeviceCallbackFP` (split from setupCallbackFP)
 
 ## Version 1.7.1 (Oct 18, 2022, JST)